modify packaging
