+++ app/src/main/java/org/sopt/sample/ui/main/gallery/GalleryFragment.kt
@@ -1,21 +1,21 @@
-package org.sopt.sample.ui.main
+package org.sopt.sample.ui.main.gallery
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import org.sopt.sample.databinding.FragmentSearchBinding
+import org.sopt.sample.databinding.FragmentGalleryBinding
 
-class SearchFragment : Fragment() {
-    private var _binding: FragmentSearchBinding? = null
+class GalleryFragment : Fragment() {
+    private var _binding: FragmentGalleryBinding? = null
     private val binding get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 바인딩 초기화 에러가 발생했습니다." }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentSearchBinding.inflate(layoutInflater)
+        _binding = FragmentGalleryBinding.inflate(layoutInflater)
         return binding.root
     }
